requirements.txt
absl-py==2.3.1
accelerate==0.30.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
attrs==25.3.0
bitsandbytes==0.43.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
datasets==2.19.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.116.1
filelock==3.13.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2024.3.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
keras==3.11.3
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml_dtypes==0.5.3
mlc-llm-nightly==0.1.dev1524
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
namex==0.1.0
networkx==3.3
numpy==2.1.2
onnx==1.19.0
onnx-tf==1.6.0
onnx2tf==1.28.2
onnxruntime==1.22.1
openai==1.102.0
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
peft==0.10.0
pillow==11.0.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
setuptools==70.2.0
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.1.0
tf2onnx==1.16.1
tf_keras==2.19.0
thefuzz==0.22.1
tiktoken==0.11.0
tokenizers==0.19.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.41.2
trl==0.8.6
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tyro==0.9.28
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
word2number==1.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
